+++ packages/signals_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 5.0.1
+
+- Switching `SignalsObserver` to use `log` instead of `print`
+- Only setting `SignalsObserver` to use `DevToolsSignalsObserver` in `kDebugMode`
+
 ## 5.0.0
 
 - Adding `previousValue`/`initialValue` to `Signal`
@@ -7,7 +12,7 @@
 - Removing old deprecated methods
 - Adding more inline code documentation
 - ListSignal/SetSignal now extends IterableSignal
-- Fixing ChnageStackSignal getters for history/redos to return an iterable list
+- Fixing ChangeStackSignal getters for history/redos to return an iterable list
 - Fixing async signal isCompleted race condition
 
 ## 4.5.0